Add multi input chain run
