+++ requirements.txt
@@ -1,3 +1,3 @@
-agixtsdk==0.0.16
+agixtsdk==0.0.17
 semantic-kernel==0.3.4.dev0
 g4f==0.0.1.3
